+++ Map/src/RMMapOverlayView.h
@@ -9,7 +9,7 @@
 #import <Cocoa/Cocoa.h>
 
 @interface RMMapOverlayView : NSView
-- (unsigned)sublayersCount;
+- (NSUInteger)sublayersCount;
 
 - (void)addSublayer:(CALayer *)aLayer;
 - (void)insertSublayer:(CALayer *)aLayer atIndex:(unsigned)index;
