+++ Map/src/RMMapOverlayView.m
@@ -37,7 +37,7 @@
     return self;
 }
 
-- (unsigned)sublayersCount
+- (NSUInteger)sublayersCount
 {
     return [self.layer.sublayers count];
 }
